+++ star/movearch.h
@@ -1,8 +1,8 @@
-/* @(#)movearch.h	1.2 03/06/12 Copyright 1993, 1995, 2001-2003 J. Schilling */
+/* @(#)movearch.h	1.3 20/07/08 Copyright 1993, 1995, 2001-2020 J. Schilling */
 /*
  *	Handle non-file type data that needs to be moved from/to the archive.
  *
- *	Copyright (c) 1993, 1995, 2001-2003 J. Schilling
+ *	Copyright (c) 1993, 1995, 2001-2020 J. Schilling
  */
 /*
  * The contents of this file are subject to the terms of the
@@ -20,15 +20,15 @@
 
 typedef struct {
 	char	*m_data;	/* Pointer to data to be moved ftom/to arch */
-	int	m_size;		/* Size of data to be moved from/to arch    */
+	size_t	m_size;		/* Size of data to be moved from/to arch    */
 	int	m_flags;	/* Flags holding different move options	    */
 } move_t;
 
 #define	MF_ADDSLASH	0x01	/* Add a slash to the data on archive	    */
 
 
-extern	int	move_from_arch	__PR((move_t *move, char *p, int amount));
-extern	int	move_to_arch	__PR((move_t *move, char *p, int amount));
+extern	ssize_t	move_from_arch	__PR((move_t *move, char *p, size_t amount));
+extern	ssize_t	move_to_arch	__PR((move_t *move, char *p, size_t amount));
 
-#define	vp_move_from_arch ((int(*)__PR((void *, char *, int)))move_from_arch)
-#define	vp_move_to_arch	((int(*)__PR((void *, char *, int)))move_to_arch)
+#define	vp_move_from_arch ((ssize_t(*)__PR((void *, char *, size_t)))move_from_arch)
+#define	vp_move_to_arch	((ssize_t(*)__PR((void *, char *, size_t)))move_to_arch)
